requirements.txt
python-dotenv
youtube-transcript-api
google-generativeai
uvicorn
fastapi